+++ src/control_plane/migration/sql_server/mod.rs
@@ -1,6 +1,8 @@
 mod backup_sql_server_database;
 mod restore_sql_server_database;
 mod sql_server_backup_options;
+mod sql_server_migration_operations;
+mod sql_server_migration_operations_options;
 mod sql_server_restore_options;
 mod sql_server_verify_target_options;
 mod verify_sql_server_target;
@@ -8,6 +10,8 @@ mod verify_sql_server_target;
 pub(crate) use backup_sql_server_database::backup_sql_server_database;
 pub(crate) use restore_sql_server_database::restore_sql_server_database;
 pub(crate) use sql_server_backup_options::SqlServerBackupOptions;
+pub(crate) use sql_server_migration_operations::SqlServerMigrationOperations;
+pub(crate) use sql_server_migration_operations_options::SqlServerMigrationOperationsOptions;
 pub(crate) use sql_server_restore_options::SqlServerRestoreOptions;
 pub(crate) use sql_server_verify_target_options::SqlServerVerifyTargetOptions;
 pub(crate) use verify_sql_server_target::verify_sql_server_target;